+++ lum-server/src/api/v1/asset-usage-agreement-restriction.js
@@ -26,7 +26,7 @@ const dbAssetUsageAgreement = require('../../db/asset-usage-agreement');
  * @param  {} next
  */
 const putAssetUsageAgreementRestriction = async (req, res, next) => {
-    utils.logInfo(res, `api putAssetUsageAgreementRestriction(${res.locals.paramKeys})`);
+    lumServer.logger.info(res, `api putAssetUsageAgreementRestriction(${res.locals.paramKeys})`);
     await pgclient.runTx(res,
         dbAssetUsageAgreement.validateAssetUsageAgreementRestriction,
         dbAssetUsageAgreement.putAssetUsageAgreementRestriction,
@@ -43,7 +43,7 @@ const putAssetUsageAgreementRestriction = async (req, res, next) => {
  * @param  {} next
  */
 const revokeAssetUsageAgreementRestriction = async (req, res, next) => {
-    utils.logInfo(res, `api revokeAssetUsageAgreementRestriction(${res.locals.paramKeys})`);
+    lumServer.logger.info(res, `api revokeAssetUsageAgreementRestriction(${res.locals.paramKeys})`);
     await pgclient.runTx(res,
         dbAssetUsageAgreement.revokeAssetUsageAgreementRestriction,
         dbAssetUsageAgreement.groomAssetUsageAgreement,