+++ lum-server/src/api/v1/asset-usage-event.js
@@ -20,6 +20,7 @@ const pgclient = require('../../db/pgclient');
 const assetUsage = require('./asset-usage');
 const dbAssetUsageReq = require('../../db/asset-usage-req');
 const dbAssetUsage = require('../../db/asset-usage');
+const acuLogger = require('../../logger-acumos');
 
 /**
  * api to get asset-usage-event from database
@@ -28,7 +29,7 @@ const dbAssetUsage = require('../../db/asset-usage');
  * @param  {} next
  */
 const getAssetUsageEvent = async (req, res, next) => {
-    utils.logInfo(res, `api getAssetUsageEvent(${res.locals.params.assetUsageId})`);
+    lumServer.logger.info(res, `api getAssetUsageEvent(${res.locals.params.assetUsageId})`);
     res.locals.dbdata.assetUsageEvent = null;
     await pgclient.runTx(res, dbAssetUsage.getAssetUsageEvent);
     if (!res.locals.dbdata.assetUsageEvent) {
@@ -39,7 +40,7 @@ const getAssetUsageEvent = async (req, res, next) => {
             response.setHttpStatus(res, res.locals.dbdata.assetUsageEvent.responseHttpCode, "assetUsageEvent");
         }
     }
-    utils.logInfo(res, `out api getAssetUsageEvent(${res.locals.params.assetUsageId})`);
+    lumServer.logger.debug(res, `out api getAssetUsageEvent(${res.locals.params.assetUsageId})`);
     next();
 };
 /**
@@ -56,7 +57,7 @@ const putAssetUsageEvent = async (req, res, next) => {
     res.locals.params.action = assetUsageEvent.action;
     res.locals.params.swTagId = assetUsageEvent.swTagId;
 
-    utils.logInfo(res, `api putAssetUsageEvent(${res.locals.params.assetUsageId}, ${res.locals.params.action})`);
+    lumServer.logger.info(res, `api putAssetUsageEvent(${res.locals.params.assetUsageId}, ${res.locals.params.action})`);
     await pgclient.runTx(res,
         dbAssetUsageReq.putAssetUsageReq,
         dbAssetUsage.putAssetUsageEvent,
@@ -72,6 +73,6 @@ const router = new Router();
 router.use(assetUsage.validateParams);
 router.route('/')
     .get(getAssetUsageEvent)
-    .put(putAssetUsageEvent);
+    .put(acuLogger.startLogForAcumos, putAssetUsageEvent);
 
 module.exports = router;