+++ lum-server/src/api/v1/asset-usage-tracking.js
@@ -35,7 +35,7 @@ const validateParams = (req, res, next) => {
  * @param  {} next
  */
 const getAssetUsageTracking = async (req, res, next) => {
-    utils.logInfo(res, `api getAssetUsageTracking(${res.locals.params.softwareLicensorId})`);
+    lumServer.logger.info(res, `api getAssetUsageTracking(${res.locals.params.softwareLicensorId})`);
     res.locals.response.title = `asset-usage-tracking for software-licensor ${res.locals.params.softwareLicensorId}`;
     res.locals.response.asOf = new Date();
     res.locals.response.softwareLicensorId = res.locals.params.softwareLicensorId;
@@ -45,7 +45,7 @@ const getAssetUsageTracking = async (req, res, next) => {
     res.locals.response.assetUsages      = res.locals.dbdata.assetUsageTracking.filter(row =>row.assetUsageType === 'assetUsage'     ).map(row => row.response);
     res.locals.response.assetUsageEvents = res.locals.dbdata.assetUsageTracking.filter(row =>row.assetUsageType === 'assetUsageEvent').map(row => row.response);
 
-    utils.logInfo(res, `out api getAssetUsageTracking(${res.locals.params.softwareLicensorId})`);
+    lumServer.logger.debug(res, `out api getAssetUsageTracking(${res.locals.params.softwareLicensorId})`);
     next();
 };